requirements.txt
discord.py==2.0.0
requests==2.28.1
pdf2image==1.16.0
Pillow==9.2.0